requirements.txt
yfinance
pandas
tqdm
datetime
psycopg2-binary
openpxl
matplotlib